[JA-14] Refactor code
